Allow usage of FITS Servlet
FITS servlet is a verison of FITS that runs on a web
server. It's much more efficient than running it via
the shell script, because it doesn't need to be reinitialized
each time it is called. That improves performance when
running running FITS on a large number of files.

This PR allows you to use that version of FITS by setting
the `FITS_SERVLET_URL` and using the `FitsServlet` Characterizer
class.